% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{reg_coef}}{numeric.}

\item{\code{fit_val}}{numeric.}

\item{\code{res_val}}{numeric.}

\item{\code{dof}}{numeric.}

\item{\code{res_var}}{matrix.}

\item{\code{var_reg_coef}}{numeric.}

\item{\code{t_val}}{numeric.}

\item{\code{p_val}}{array.}
}}


