% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\value{
Linear Regression
}
\description{
Linear Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{reg_coef}}{array.}

\item{\code{fit_val}}{numeric.}

\item{\code{res_val}}{numeric.}

\item{\code{dof}}{numeric.}

\item{\code{res_var}}{matrix.}

\item{\code{var_reg_coef}}{numeric.}

\item{\code{t_val}}{numeric.}

\item{\code{p_val}}{array.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{This function returns regression coefficients}

\item{\code{pred()}}{This function returns fitted value}

\item{\code{print()}}{This function prints out the coefficients and coefficient names}

\item{\code{resid()}}{This function returns residuals value}

\item{\code{summary()}}{This function prints out the coefficients with their standard error, t-value and p-value}
}}

\examples{
data(iris)
mod_object <- linreg(Petal.Length~Species, data = iris)
print(mod_object)


}
