% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression}
\arguments{
\item{formula}{object of class 'ridgereg': a symbolic description of the model to be fitted.}

\item{data}{data frame containing the variables in the model.}

\item{lambda}{a numeric data, which is a ridge constant.}
}
\value{
Returns an object of class 'ridgereg'.
}
\description{
A package that take a formula object as well as a dataset and return a ridgereg object.
Implement a RC class to handle special functions: print(), coef() and predict().
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{lambda}}{numeric, ridge constant.}

\item{\code{X}}{matrix which contains independent values.}

\item{\code{y}}{matrix which contains dependent values.}

\item{\code{Q}}{matrix, from the QR decomposition.}

\item{\code{R}}{upper triangular matrix.}

\item{\code{Qy}}{matrix.}

\item{\code{reg_coef}}{Regression coefficients.}

\item{\code{fit_val}}{The fitted values.}

\item{\code{res_val}}{The residuals.}

\item{\code{dof}}{The degree of freedom.}

\item{\code{res_var}}{The residual variance.}

\item{\code{var_reg_coef}}{The variance of the regression coefficients.}

\item{\code{t_val}}{The t-values for each coefficient.}

\item{\code{p_val}}{P-values for each regression coefficients.}

\item{\code{data_name}}{character.}

\item{\code{X_mean}}{numeric.}

\item{\code{X_var}}{numeric.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{This function returns regression coefficients}

\item{\code{print()}}{This function prints out the coefficients and coefficient names}
}}

\examples{
data(iris)
mod_object <- ridgereg(Petal.Length~Species, data = iris, lambda = 0.5)
mod_object$print()
mod_object$coef()
mod_object$predict()

}
