% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Airport Delay Visualize}
\usage{
visualize_airport_delays()
}
\value{
Return plot
}
\description{
Function creates a plot that visualize the mean delay of flights
for different airports by longitude and latitude using ggplot2 library.
}
\examples{
visualize_airport_delays()

}
