% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression}
\arguments{
\item{formula}{object of class 'linreg': a symbolic description of the model to be fitted.}

\item{data}{data frame containing the variables in the model.}
}
\value{
Returns an object of class 'linreg'.
}
\description{
A package to handle linear regression models with using linear algebra.
Implement a RC class to handle special functions: print(), plot(), resid(), 
pred(), coef() and summary().
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{reg_coef}}{Regression coefficients.}

\item{\code{fit_val}}{The fitted values.}

\item{\code{res_val}}{The residuals.}

\item{\code{dof}}{The degree of freedom.}

\item{\code{res_var}}{The residual variance.}

\item{\code{var_reg_coef}}{The variance of the regression coefficients.}

\item{\code{t_val}}{The t-values for each coefficient.}

\item{\code{p_val}}{P-values for each regression coefficients.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{This function returns regression coefficients}

\item{\code{pred()}}{This function returns fitted value}

\item{\code{print()}}{This function prints out the coefficients and coefficient names}

\item{\code{resid()}}{This function returns residuals value}

\item{\code{summary()}}{This function prints out the coefficients with their standard error, t-value and p-value}
}}

\examples{
data(iris)
mod_object <- linreg(Petal.Length~Species, data = iris)
mod_object$print()
mod_object$resid()
mod_object$pred()
mod_object$coef()
mod_object$summary()

}
